Velocitat i flip de player

+++ LoaderParams.h
@@ -6,7 +6,7 @@
 class LoaderParams
 {
 public:
-	LoaderParams(int x, int y, int width, int height, 
+	LoaderParams(float x, float y, int width, int height,
 		const char* textureID, int currentRow, int sprits, int flip){
 		m_x = x;
 		m_y = y;
@@ -16,21 +16,23 @@ public:
 		m_currentRow = currentRow;
 		m_num = sprits;
 	}
-	int getX() const { return m_x; }
-	int getY() const { return m_y; }
+	float getX() const { return m_x; }
+	float getY() const { return m_y; }
 	int getWidth() const { return m_width; }
 	int getHeight() const { return m_height; }
 	int getCurrentRow() const { return m_currentRow; }
 	const char* getTextureID() const { return m_textureID; }
 	int getSprits() const { return m_num; }
+	
 private:
-	int m_x;
-	int m_y;
+	float m_x;
+	float m_y;
 	int m_width;
 	int m_height;
 	int m_currentRow;
 	const char* m_textureID;
 	int m_num;
+	
 };
 
 #endif LOADERPARAMS_H
